refactor(toolkit-lib): remove shared-private.ts and update references
