--- packages/@aws-cdk/toolkit-lib/lib/api/shared-private.ts
@@ -1,18 +1,0 @@
-/* eslint-disable import/no-restricted-paths */
-
-// Polyfills first
-import '../private/dispose-polyfill';
-
-// private code
-export * from '../private';
-
-// private apis
-export * from './io/private';
-export * from './aws-auth/private';
-export * from './cloud-assembly/private';
-
-export * as cfnApi from './deployments/cfn-api';
-export { makeRequestHandler } from './aws-auth/awscli-compatible';
-
-// Context Providers
-export * as contextproviders from '../context-providers';
