+++ packages/@aws-cdk/toolkit-lib/test/toolkit/base-credentials.test.ts
@@ -1,5 +1,5 @@
+import { SdkProvider } from '../../lib/api/aws-auth/private';
 import { BaseCredentials } from '../../lib/api/aws-auth/types';
-import { SdkProvider } from '../../lib/api/shared-private';
 import { Toolkit } from '../../lib/toolkit/toolkit';
 import { appFixture, TestIoHost } from '../_helpers';
 import { MockSdk } from '../_helpers/mock-sdk';
